Photos of ETL Steps before Debugging
Use all the steps in order for each package to edit them so that they can upload to the database correctly.

Change the connection name for the file.

Browse for the file.

Go to the Advanced tab and click on Suggested Types. Click OK.

Edit the Database flow by typing in your server name and the AssistanceTech database.
Example: <SERVER_NAME>.AssitanceTech

Lastly, Edit the table name for that package. If the package is Customers, then the table being made named should be called Customers.

Debug the package and It should appear in the database.